+++ xenon-common/src/main/java/com/vmware/xenon/services/common/AuthCredentialsService.java
@@ -24,6 +24,7 @@ import com.vmware.xenon.common.Service;
 import com.vmware.xenon.common.ServiceDocument;
 import com.vmware.xenon.common.ServiceDocumentDescription;
 import com.vmware.xenon.common.StatefulService;
+import com.vmware.xenon.common.Utils;
 
 /**
  * Describes the authentication credentials to authenticate with internal/external APIs.
@@ -98,10 +99,20 @@ public class AuthCredentialsService extends StatefulService {
 
     @Override
     public void handlePatch(Operation patch) {
+        if (!patch.hasBody()) {
+            patch.fail(new IllegalArgumentException("body is required"));
+            return;
+        }
         AuthCredentialsServiceState currentState = getState(patch);
-        AuthCredentialsServiceState patchBody = patch.getBody(AuthCredentialsServiceState.class);
-
-        updateState(patchBody, currentState);
+        try {
+            if (!Utils.mergeWithState(currentState, patch)) {
+                AuthCredentialsServiceState patchBody = patch.getBody(AuthCredentialsServiceState.class);
+                updateState(patchBody, currentState);
+            }
+        } catch (NoSuchFieldException | IllegalAccessException e) {
+            patch.fail(e);
+            return;
+        }
         patch.setBody(currentState).complete();
     }